backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
gunicorn==23.0.0
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
sqlparse==0.4.4
typing_extensions==4.11.0
python-dotenv==1.0.0
django-allauth==65.4.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.162.0
